% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{arrrange}
\alias{arrrange}
\title{Populate a numeric range into discrete sequence}
\usage{
arrrange(range, step, simplify = FALSE)
}
\arguments{
\item{range}{A NumRange object, a valid character vector can be converted into NumRange object, for convenience.}

\item{step}{A vector of steps for each member range. Default is each range's (max-min)/10. Length of step should be 1 or number of member ranges.}

\item{simplify}{Default is FALSE. TRUE for automatically unlist the result vector.}
}
\value{
A list of sequence if simiplify = FALSE or a sequence otherwise.
}
\description{
A wrapper for R6 method NumRange$arrrange
}
