% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRange.R
\name{as_.NumRange}
\alias{as_.NumRange}
\alias{as.NumRange}
\title{Convert a character string to object NumRange.}
\usage{
as_.NumRange(x)
}
\arguments{
\item{x}{An vector of type character.
x should be in the form of bracketed string, leaded by "[" or "(", and ended with ")" or "]".

Other form won't be accepted.}
}
\description{
Convert a character string to object NumRange.
}
