% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{as_set}
\alias{as_set}
\alias{as.set}
\title{Set Generator}
\usage{
as_set(..., .list)
}
\arguments{
\item{...}{a spreaded list of values.}

\item{.list}{an unspreaded list of values.}
}
\description{
A wrapper of .Set$new()
}
