% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRange.R
\name{is_.NumRange}
\alias{is_.NumRange}
\alias{is.NumRange}
\title{Check if an object is of class NumRange.}
\usage{
is_.NumRange(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
A logical vector of length==length(x)
}
\description{
Check if an object is of class NumRange.
}
