% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_}
\alias{is_}
\title{Check if an object is of a class}
\usage{
is_(x, class)
}
\arguments{
\item{x}{an object}

\item{class}{a vector of class. A value of FALSE, TRUE, and NULL will trigger respective vectorized is_ functions.}
}
\description{
Check if an object is of a class
}
