% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_character}
\alias{is_character}
\title{Element-wise wrapper for is.character}
\usage{
is_character(x)
}
\arguments{
\item{x}{a vector or object}
}
\value{
a logical vector
}
\description{
Element-wise wrapper for is.character
}
