% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{is_covered_by}
\alias{is_covered_by}
\alias{is.covered_by}
\title{Check if a numeric range is covered within another one}
\usage{
is_covered_by(range, x)

is_covered_by(range, x)
}
\arguments{
\item{range}{A NumRange object, a valid character vector can be converted into NumRange object, for convenience.}

\item{x}{A vector or object of R6class NumRange to be check, a valid character vector can be converted into NumRange object, for convenience.}

\item{x}{A vector or object of R6class NumRange to be check, a valid character vector can be converted into NumRange object, for convenience.}

\item{range}{A NumRange object, a valid character vector can be converted into NumRange object, for convenience.}
}
\value{
A logical vector

A logical vector
}
\description{
A wrapper for R6 method NumRange$covered_by

A wrapper for R6 method NumRange$cover
}
