% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_falsified}
\alias{is_falsified}
\title{Check if some objects don't return a legible value}
\usage{
is_falsified(x)
}
\arguments{
\item{x}{a vector or object}
}
\value{
a logical vector
}
\description{
Check if some objects don't return a legible value
}
