% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{is_identicalRange}
\alias{is_identicalRange}
\alias{is.identicalRange}
\title{Check if two numeric ranges are identical}
\usage{
is_identicalRange(x, y)
}
\arguments{
\item{x, y}{two numeric ranges of R6Class 'NumRange'}
}
\value{
a logical value
}
\description{
Check if two numeric ranges are identical
}
