% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{is_NullRange}
\alias{is_NullRange}
\alias{is.NullRange}
\title{Check if a numeric range is empty}
\usage{
is_NullRange(x)
}
\arguments{
\item{x}{a numeric range of R6Class 'NumRange'}
}
\value{
a logical value
}
\description{
Check if a numeric range is empty
}
