% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{is_NumRange}
\alias{is_NumRange}
\alias{is.cNumRange}
\title{Check if an object is of class cNumRange.}
\usage{
is_NumRange(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
A logical vector of length==length(x)
}
\description{
Check if an object is of class cNumRange.
}
