% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{NR_antiunion}
\alias{NR_antiunion}
\title{Anti union 2 NumRange objects}
\usage{
NR_antiunion(x, y)
}
\arguments{
\item{x, y}{Two NumRange objects to be anti-unioned}
}
\value{
an object of R6Class 'NumRange'
}
\description{
A wrapper for R6 method NumRange$antiunion_with
}
