% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{NR_intersect}
\alias{NR_intersect}
\title{Intersect of 2 NumRange objects}
\usage{
NR_intersect(x, y)
}
\arguments{
\item{x, y}{Two NumRange objects to be intersected}
}
\value{
an object of R6Class 'NumRange'
}
\description{
A wrapper for R6 method NumRange$intersect_with
}
