% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\name{NR_union}
\alias{NR_union}
\title{Union 2 NumRange objects}
\usage{
NR_union(x, y)
}
\arguments{
\item{x, y}{Two NumRange objects to be unioned.}
}
\value{
an object of R6Class 'NumRange'
}
\description{
A wrapper to union 2 NumRange object, which is in fact a wrapper for R6 method $union_with
}
