% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNumRange.R
\docType{data}
\name{NumRange}
\alias{NumRange}
\title{Combined Continuous Numeric Range}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cNumRange
 $new(..., list)
 $clone()
 $include(x)
 $cut(..., mode)
 $extract(id)
 $del(range)
 $print()
}
\arguments{
\item{...}{For method 'new': spreaded list of NumRange.
For method 'cut': spreaded list of split points}

\item{id}{the number of element NumRange within the object. If left missing, the method will return all elements.}

\item{list}{A list of NumRange. Used in case NumRanges are in a form of list.}

\item{mode}{Mode for returned cutted NumRanges, if missing, defaults of element NumRanges are used.}

\item{range}{A range that will be deleted from the host range.}

\item{x}{A numeric vector to check for inclusion.}
}
\value{
An object of R6Class "cNumRange"
}
\description{
An R6 object for combined continuous numeric ranges
}
\details{
This class is used when simple NumRange is not enough.
It combines many NumRange into one, thus form a better constraint for complex data.
You can take back the simple NumRanges as its elements by calling the extract method.

Please refer to NumRange for more details.
}
\keyword{datasets}
