% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\docType{data}
\name{Set}
\alias{Set}
\title{Set class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Set
 $new(..., .list)
 $print()
 $include(x)
 $raw
 $val
}
\arguments{
\item{...}{a spreaded list of values.}

\item{.list}{an unspreaded list of values.}

\item{x}{a vector of objects to check if they are included in the set}
}
\value{
$new: an object of R6 class "Set"

$include: a logical vector of length == length(x)

$val: a list

$raw: a character string

Note that $print, $new invisibly return itself and support method chaining.
}
\description{
An R6 object for discrete value set
}
\section{Class Methods}{

$new: the class constructor

$print: the print method

$include: check if an object is included in the set

$raw: print a beautified string describing the set

$val: return a conventional list
}

\keyword{datasets}
