% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRange.R
\name{union..NumRange}
\alias{union..NumRange}
\title{Union 2 NumRange objects}
\usage{
\method{union}{.NumRange}(x, y)
}
\arguments{
\item{x, y}{Two NumRange objects to be unioned.}
}
\description{
A S3 method to union 2 NumRange object, which is in fact a wrapper for R6 method $union_with
}
