% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumRange.R
\name{union.default}
\alias{union.default}
\title{Union 2 objects}
\usage{
\method{union}{default}(x, y)
}
\arguments{
\item{x, y}{Two objects to be unioned.}
}
\description{
A function to union 2 object, now support method and NumRange
}
